switch config from object to array of arrays
Objects does not define an order for their properties. So with
"mapping": {
  "v2/00-init": "v2/01-adjust-generated-code",
  "v2/01-adjust-generated-code": "v2/02-components-data-binding"
}
it is not clear if we should merge "v2/00-init" to "v2/01-adjust-generated-code" first and then "v2/01-adjust-generated-code" to "v2/02-components-data-binding" or vice versa.
But we need exactly that order. Because of this the config should be an array of arrays, to always be deterministic and exactly know which branch should go into which and in which order.

Also an array is easier to maintain. You don't have to repeat yourself. The branch to merge to ist the next in the array.

We use an array of arrays instead of a one flat array to be expressive as the object notation. The first branch name in an inner array specifies the start point. The other branches in the same array get merged in exactly that order. Multiple inner arrays defines independent start-merge-points.